feat: 출석 체크 기능 추가 문제 풀이를 통한 출석 체크 및 포인트 지급 기능을 구현했습니다. controller, entity, repository, service 파일을 통해 기능이 구현되도록 하였습니다.

+++ src/main/java/com/ssafy/solvedpick/attendance/domain/AttendanceRecord.java
@@ -1,6 +1,5 @@
 package com.ssafy.solvedpick.attendance.domain;
 
-import com.ssafy.solvedpick.members.domain.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -24,7 +23,7 @@ public class AttendanceRecord {
     private Long id;
 
     @Column(name = "user_id", nullable = false)
-    private Integer userId; 
+    private Long userId; 
 
     @Column(name = "attendance_date", nullable = false)
     private LocalDate attendanceDate;
@@ -33,12 +32,12 @@ public class AttendanceRecord {
     @Column(name = "created_at", updatable = false)
     private LocalDateTime createdAt;
 
-    public AttendanceRecord(Integer userId) {
+    public AttendanceRecord(Long userId) {
         this.userId = userId;
         this.attendanceDate = LocalDate.now();
     }
 
-    public static AttendanceRecord create(Integer userId) {
+    public static AttendanceRecord create(Long userId) {
         return new AttendanceRecord(userId);
     }
 }
